Add adress, item, order, category controllers and fix few bugs at backend, add basic service to frontend

+++ Backend/src/main/java/project/onlineshop/service/interfaces/ItemService.java
@@ -3,11 +3,13 @@ package project.onlineshop.service.interfaces;
 import org.springframework.web.multipart.MultipartFile;
 import project.onlineshop.dto.Response;
 
+import java.math.BigDecimal;
+
 public interface ItemService {
 
-    Response createItem(Long categoryId, MultipartFile image, String name, String description, Double price);
+    Response createItem(Long categoryId, MultipartFile image, String name, String description, BigDecimal price);
 
-    Response updateItem(Long itemId, Long categoryId, MultipartFile image, String name, String description, Double price);
+    Response updateItem(Long itemId, Long categoryId, MultipartFile image, String name, String description, BigDecimal price);
 
     Response deleteItem(Long itemId);
